Passbild und Signatur werden nicht merh in der DB gespeichert.

+++ app/Http/Controllers/ApplicationController.php
@@ -57,18 +57,14 @@ class ApplicationController extends Controller
         //Falls keine Signatur hinterlegt ist, wurde sie direkt beim Drucken hochgeladen
         //Erstelle ine temporäre Datei für die Signatur; sie wird gelöscht, sobald das PDF-Dokument vorliegt
         //nötig, das Mpdf sehr lange html-strings (vgl. blob) nicht verträt
-
-        if (is_null($help->signature)) {
-            //$format['signature'] = base64_encode(file_get_contents($request->file('signature')));
-            //$format['sig_datatype'] = $request->file('signature')->extension();
+        if($request->signature) {
+            //$format['signature'] = base64_encode($help->signature);
+            //$format['sig_datatype'] = $help->sig_datatype;
             $filename = uniqid() . '.' . $request->file('signature')->extension();
             file_put_contents(storage_path('app/temp/' . $filename), file_get_contents($request->file('signature')));
         }
         else {
-            //$format['signature'] = base64_encode($help->signature);
-            //$format['sig_datatype'] = $help->sig_datatype;
-            $filename = uniqid() . '.' . $help->sig_datatype;
-            file_put_contents(storage_path('app/temp/' . $filename), $help->signature);
+            $filename = false;
         }
         $format['signature'] = $filename;
 
@@ -93,7 +89,9 @@ class ApplicationController extends Controller
 </table>');
 
         $pdf->WriteHTML(view("bewerbungen.applications.pdfNew", compact("application", "content", "resume", "format"))->render());
-        unlink(storage_path('app/temp/' . $filename));
+        if ($filename) {
+            unlink(storage_path('app/temp/' . $filename));
+        }
         return $pdf->Output("Bewerbungsanschreiben.pdf", 'I');
         //return view("bewerbungen.applications.pdfNew");
         //return view("bewerbungen.applications.pdfNew", compact("application", "content", "resume", "format"));
@@ -108,15 +106,8 @@ class ApplicationController extends Controller
     public function edit(Application $application)
     {
         $user = $application->user;
-        $resume = $user->resume;
-        if (!(is_null($resume) || is_null($resume->signature))) {
-            $signature = base64_encode($resume->signature);
-        }
-        else{
-            $signature = false;
-        }
 
-        return view("bewerbungen.applications.edit", compact("application", "signature"));
+        return view("bewerbungen.applications.edit", compact("application"));
     }
 
     /**
